untested recomMethod
This method has not been tested due to it not being implemented quite
yet in the application. This should be the basis for the method though
unless we figure out a more complex system to determine recommendations.
Planning on adding the feature to call this method tomorow during our
work time (wanted to avoid conflicts)