finished some stuff
still a few errors in redis and mongo, inconsistent data left in
databases due to these errors. Errors are, in RedisCatchup- if error is
thrown log is still being updated as complete. in MongoCatchup- prof
class pair is never being deleted. Obvious inconsistency is if you
create a class for a professor, delete the class, then search for the
class two different ways (get proffessor and search class)